Add SB_MODULAR_BUILD to starboard/configuration.h
Pass SbEventHAndle as a function pointer to SbRunStarboardMain. This
breaks Starboard's dependecy on Cobalt allowing Starboard to be built as
a standalone shared library.

SbRunStarboardMain is defined as the entry method into starboard with the client application event handler.

Update all platform Applications class to pass the SbEventHandle pointer to
starboard::shared::Application

SbEventHandle will be visible to the loader_app via SB_EXPORT_PLATFORM.
SB_IMPORT SbEventHandle is removed as the standalone starboard dynamic library doesn't need to import any symbols as it's not
dependent on any app.

b/265874860

Change-Id: I00bd093a50c48e7dc79ba9c875e5dbfe902fbc3f
NOKEYCHECK=True
GitOrigin-RevId: 9e55b33916f5252b7bd5c0cea364e0ab0605cf12