Implement Crashpad handler for AOSP Evergreen
This change builds a native build of Crashpad handler and packages it in
the APK so that it can be run after a crash. It also has the Cobalt
process extract the CA root certificates from the APK to the file system
so that the Crashpad handler process can access them. Please see
***REMOVED*** for more background.

b/256660693

Change-Id: Ic0a01eb82cd2d8b19cbfcb49c714543b5a5e1e2c
NOKEYCHECK=True
GitOrigin-RevId: 8a208451be9fe72ec33c8b7b1689014725868dc9