Fix potential crash in FetcherCache
FetcherCache caches a resource on loading finish.  When two resources
with the same url are being loaded concurrently, it's possible that the
size of the resource will be counted twice but only one of them will be
added to the cache.  This leads to potential crash when removing entries
from cache.

This is now fixed by no longer double counting the resource size in such
case.

b/270993319

Change-Id: If0723c92526123ced60121e82f4cdd71a4022d44
NOKEYCHECK=True
GitOrigin-RevId: f746e8769524a50de4bbc336e4f725d0cc02e6b6