[media] Optimize Opus audio decoder for batch sample writing
To improve the efficiency of the Opus audio decoder, we propose
implementing a non-blocking approach for writing batches of packets, by
scheduling the write operation for each packet separately.

This will allow the PlayerWorker thread to accept other tasks while
decoding each packet, instead of blocking on a batch write operation. By
using this approach, we aim to reduce the overall latency of the decoder
and increase its throughput, while also minimizing the impact on the
system's resources.

The experimental results indicate that when the batch size is set to 8,
the response time can be reduced from 6.107(milliseconds in average):
1. to 3.194 when write 1 buffers per schedule.
2. to 3.613 when write 2 buffers per schedule.

b/264302691

Change-Id: I0525f3ef80c2701fe63c875f9aa5abdf2513d7fa
(cherry picked from commit dd8aac12390156e2d2223d2797581a0d5cca114a)
NOKEYCHECK=True
GitOrigin-RevId: a28ec609ca4c309dfbd0b8b05644d358a26f82b4