Add CSP for worker script fetching, and certificates errors in wpt.
This adds using of the csp delegate when fetching worker scripts, and
sets ignore_certificate_errors to allow running tests that use https.

After this, service-worker/service-worker-csp-script.https.html has a
final issue causing a failure because it unexpectedly fails a CORS check
due to a missing Access-Control-Allow-Origin, as well as flakyness due
to the still missing b/234788479.

The OnCspPolicyChanged callback is moved from WebModule to
WindowOrWorkerGlobalScope where it can access the CspDelegate and the
GlobalEnvironment as needed, and where it will also work for workers.

This also moves all CSP options to a struct CspDelegate::Option, with
member naming that is consistent with the type names, and some related
cleanups that make using CspDelegateFactory::Create() a lot simpler,
and moves the StatTracker to the WindowOrWorkerGlobalScope::Options so
that it doesn't have to be forwared everywhere separately.

b/234788479
b/225037465
b/241801523

Change-Id: Iec80b43cbcc0940cf473e1caf25c3f32af76be4f
NOKEYCHECK=True
GitOrigin-RevId: 7931a4111847214b992bb4ce2a6d23e703c9c325