Abort any remaining service worker on app exit.
This ensures that all still active service workers get an Abort call
when the ServiceWorkerRegistry is destructed on application shutdown.

Also, make some defensive changes such as switching to a TimedWait so
that this destructor can never be stuck forever, because application
shutdown waits for this destructor.

This fixes
b/274832572

Change-Id: Iefbdbf35e0e55463e8b5d5a94fe643d72589db9d
NOKEYCHECK=True
GitOrigin-RevId: 638947fae4d4b68e0337aa78fd77dea8b6b2975f