Fix Python 3.11.2 installer bug in Docker Windows
The latest version of python in chocolatey is 3.11.2 which has a
regression that makes it ignore install arguments such as TargetDir.
This breaks the docker build since the Python installation directory is
no longer C:\Python3.

As a workaround, we pin the version to 3.11.0, which address the issue.
Ideally we would remove the pinned version if it is addressed at a later
time.

b/276772883

Change-Id: I211f4d6212b56fee13f616498e635e1a64a26370