[NX] The crash handler implementation was added to the project
To add a message to a crash report Dying Messages memory region is used.
The message is copied to the memory region and if a crash occurs the
memory region will be added to the crash report.

b/249762037

Change-Id: I9e32886010980bc0c055af80d471d1c5fca8a1b7
(cherry picked from commit a1ce402d9cdd3446371504c903adef6a74e42360)
NOKEYCHECK=True
GitOrigin-RevId: c7a8942089f50c9430a1cc49b2eb9c5efdf97cc0