[media] Add SbPlayer tests with drm system
This cl adds SbPlayer tests with drm system. As encrypted playback
requires key exchanges, which is not available for testing, we create
SbPlayers with supported drm systems and feed them with non-encrypted
samples, to increase our test coverage for encrypted pipelines.

This cl also add new cases for playbacks with both audio and video
streams to increase test coverage.

b/274493767