some WPT flakyness fixes (#124)
* WPT Use get_newest_worker instead of registration.installing.

This replaces a registration.installing with a
get_newest_worker(registration) as a temporary workaround for the
KI of worker installation not waiting for tasks to complete.

b/234788479

* Fix for a leak and certain flakyness in web_platform_tests.

This ensures that the ServiceWorkerRegistry doesn't leak when running
web platform tests.

This also removes the cache::Cache singleton being destroyed from
ServiceWorkerPersistentSettings.

This fixes
b/264322440
b/278764092