Turn ANRs for offloaded SourceBuffer to crash
This CL affects all platforms.

ANRs were observed when SourceBuffer offload is enabled on Android TV.
It's likely that one of the relevant threads is deadlocked during
suspend, but we are unable to get a callstack of the offending thread.

This CL turns the ANRs into crashes, by adding CHECKs around the only
mutex introduced by SourceBuffer offload, to crash when the mutex is
re-acquired on the same thread, or when it takes too long to acquire.
This doesn't resolve the original bug, but will allow us to get the
callstack of the offending thread for further debugging.

b/223283615
b/271167555