Introduce libcast_starboard_api.so.
This library contains the Starboard platform, along with Cast-specific
calls for initialization. It is consumed by Cast libraries so that vendors
may use their Starboard implementation to run Cast, rather than needing to
separately integrate their platform with Cast.

A reference implementation along with basic testing are provided to ease
the library implementation.

This change also introduces the "sb_enable_cast_api" to the build, which
is required for the Cast library to be available. This eliminates the
impact on other Cobalt code, which would otherwise need refactoring of
the "final_executable_type" and "main=StarboardMain" logic.