Temporary workaround for "subtree" checkout errors on Windows
This switches Windows workers, which are not yet ephemeral to use
a patched version of checkouts.

This guarantees always cleaning up repo state even in case of
previous bad submodules-like repo operations done by a previous
run.

In the long run, we should be using ephemeral workers, and switch
to upstream fixed version if the PR gets merged upstream.

b/282361986