[loader] Add debugging code for FetcherCache crash
The CHECK for consistency in linked_hash_map::erase() fails in
production, where we are unable to find an obvious cause.

Now FetcherCache is ref counted to ensure that there isn't any life time
issue, and extra CHECKs are added so we can have more information about
the crash once released.

b/270993319