[android] Support decode to texture for Shorts
There was a bug in the calculation of the dimensions for decode target
under decode to texture mode:

The dimensions should be represented by the texture size and the crop
rectangle of the frame in the texture.  The old code incorrectly
calculated the texture size using the crop rectangle, and then try to
calculate a "crop rectangle" out of the crop rectangle.

The code happened to work for most of the 16:9 videos, as the double
cropped rectangle keeps the right ratio (under the wrong dimensions),
but it works as the frame is rendered as a texture using relative
coordinate.

The code broke on non 16:9 videos, especially vertical videos like
Shorts.

Also did the following cleanups:
1. The `width` and `height` passed down from VideoDecoder to
   MediaCodecBridge are hint only, and used to be set to the width and
   height of the video window, which is an unnecessary dependency.
   Now they are set to the resolution of the initial video stream info.
2. Remove adaptive playback related variables from MediaCodecBridge, as
   they are always set to true.

b/261922568