Add args for launching blackboxtests for specific test set
Adds arguments for launching individual test suites, specifically
to be able to use Proxy for WPT tests that require it, but not for
others.

Also simplify argument passing and variable naming a bit.

This refactor should initially be a no-op, as long as new arguments
are not being passed by invoker, i.e. tests default to "all".

b/275077120