Move SbPlayerPipeline declaration to .h file
The declaration of SbPlayerPipeline was in sbplayer_pipeline.cc so other
code won't depend on SbPlayerPipeline directly.  Instead, they'd create
it using Pipeline::Create() (which creates an SbPlayerPipeline
internally) and access it via the Pipeline interface.

We are trying to reduce the size of sbplayer_pipeline.cc (it's >1700
loc as of today).  Moving the declaration to an .h file reduces ~400
lines, which outweighs the benefit mentioned above.

There are no functional changes, except that Pipeline::Create() is
removed, as now SbPlayerPipeline can be created directly.

b/285639635