Allow sending of browser logs to devtools.
This adds a debug_log() console command that can turn on sending browser
logs to attached debuggers.

This also adds related buttons on the Cobalt page in devtools.

This also ensures that the DebugModule and RenderOverlay are only enabled
when they are needed (lazily), so that they don't unnecessarily use
resources when no debugger is attached.

b/251825168