Allow sending of browser logs to devtools. (#562)
Updated to allow instantiating a debugger from the debugger client.

This adds a debug_log() console command that can turn on sending browser
logs to attached debuggers.

This also adds related buttons on the Cobalt page in devtools.

This also ensures that the DebugModule and RenderOverlay are only enabled
when they are needed (lazily), so that they don't unnecessarily use
resources when no debugger is attached.

b/251825168

This reverts commit 230ee2eb55644c0ccca0f661774778d533c2b096.