Delete crashing CobaltMetricsLogUploader test
This test what might happen if a non-gzipped payload gets sent by Chromium to
our libraries. This payload being compressed is fundamental to the API provided
by the underlying Chromium libraries and will not change. If it does, that's a
wider refactor and would break a lot of stuff. It's kind of just testing the
public API stays the same of //components/metrics, which really isn't a good
unit test.

b/287972956

Change-Id: I3690b93ac8fd3a26742ae41a6fca931c29cf9c5e