Fix threading-related crash in Cobalt telemetry.
The Chromium metrics libraries require any calls to their public APIs to
be invoked on the same thread in which they were created. Similarly,
H5vcc JS callback need to be invoked on the V8 thread.

This was being violated as the H5vcc callbacks were being invoked by a
non-V8 thread and H5vccMetrics was trying to call metrics APIs from a V8
thread.

The fix is to store the task runners for the
CobaltMetricsServicesManager and H5vccMetrics instances and use those to
run any tasks that must be run in the targeted thread.

This approach required a slight refactoring to allow the top level
CobaltMetricsServicesManager to provide its own public API to interact
with the underlying metrics client on a targeted thread.

I also found this approach allowed me to just pass around a
RepeatingCallback for the upload handler and remove the abstract class
that was being used before (CobaltH5vccMetricsUploaderCallback).

Added demo to verify everything is working.

b/288252273

Change-Id: I8243b27a1853c65ac9d4fe11adc95a7f4da9ed51