Change device output path on Android
GetDeviceOutputPath() currently returns the cache folder which is mapped
to kSbSystemPathCacheDirectory. According to the documentation this is
used as a local file cache. The log folder on the other hand is mapped to
kSbSystemPathDebugOutputDirectory which is to be used for "debug output
(e.g. logs, trace output, screenshots)". This matches better with the
purpose of GetDeviceOutputPath().

b/289417243