Add support to verify a Crx from a string (#780)
With in-memory updates the Crx package will still be in memory when it
needs be verified.

This change duplicates a lot of Chromium code but this should only be
temporary - please see the comments in crx_verifier.cc for more on this.

b/158043520

(cherry picked from commit b62c600a18eb4a5e2dd9522afd0e9836a7f24398)