[android] Reset frame rate estimator when the stream fps changed
When playing a stream, the key operating rate is set to MediaCodec for
resource planning. It's calculated by the playback rate and the
estimated fps. While the latter is calculated by the number of
rendered frames in duration. When the stream fps is changed, the
estimated fps will be affected by the frames already rendered with the
old fps and slowly adjusted to the new fps. This may lead to
insufficient resources prepared for the stream and cause some
performance issues.