[android] Adjust media time during 5.1 playback
Adjusts the playback time reported by GetCurrentMediaTime() to respect
the offset between the seek time and the timestamp of the first written
audio packet. As passthrough audio packets cannot be adjusted to the
seek time, this avoids an issue where the audio timestamp and the seek
time are not aligned.

As an (E)AC3 audio packet timestamp may be up to ~32 ms earlier than the
seek time, the VideoRenderAlgorithm |kBufferTooLateThreshold| is
adjusted to ensure that frames aren't dropped too early.

b/243588665

Change-Id: Ic574b6e4507fda227199b77df4dda175969ba589