Add blocking toggle to SetPersistentSettings. (#896)
Add blocking toggle to SetPersistentSettings to wait for
CommitPendingWrite() to complete. Greatly improves reliability and
accuracy of persistent_settings_test especially now that sleep can be
removed. Added blocking toggle to RemovePersistentSettings as well.

Replaces ASSERT with EXPECT when inside of closures that depend on
test_done.Signal() to avoid deadlocking.

b/280430510
b/283529011