[android] Adjust media time during 5.1 playback (#899)
* [android] Adjust media time during 5.1 playback

Adjusts the playback time reported by GetCurrentMediaTime() to respect
the offset between the seek time and the timestamp of the first written
audio packet. As passthrough audio packets cannot be adjusted to the
seek time, this avoids an issue where the audio timestamp and the seek
time are not aligned.

As an (E)AC3 audio packet timestamp may be up to ~32 ms earlier than the
seek time, the VideoRenderAlgorithm |kBufferTooLateThreshold| is
adjusted to ensure that frames aren't dropped too early.

b/243588665

(cherry picked from commit 88dc47508d146d52527511dc5d93b6addcb43064)