Remove overriden readelf assignment (typo)
This change cleans up a typo that's been around since the GN migration.

The ar program should be unaffected since the assignment on the
following line was overriding the assignment.

readelf should also be unaffected as the default ("readelf") from the
gcc_toolchain template will continue to be used.

b/185522715

Change-Id: I6b35f4ec8420d53f0fddb7d71c0fe3e885215b91