[android] Support decode to texture for Shorts (#963)
There was a bug in the calculation of the dimensions for decode target
under decode to texture mode:

The dimensions should be represented by the texture size and the crop
rectangle of the frame in the texture.  The old code incorrectly
calculated the texture size using the crop rectangle, and then try to
calculate a "crop rectangle" out of the crop rectangle.

The code happened to work for most of the 16:9 videos, as the double
cropped rectangle keeps the right ratio (under the wrong dimensions),
but it works as the frame is rendered as a texture using relative
coordinate.

The code broke on non 16:9 videos, especially vertical videos like
Shorts, and is now fixed.

Also did the following cleanups:
1. The `width` and `height` passed down from VideoDecoder to
   MediaCodecBridge are hint only, and used to be set to the width and
   height of the video window, which is an unnecessary dependency.
   Now they are set to the resolution of the initial video stream info.
2. Remove adaptive playback related variables from MediaCodecBridge, as
   they are always set to true.
3. Moved nplb/drm_helpers.* to starboard/shared/starboard/drm/ to they
   can be used by starboard_platform_tests.
4. Refine `SbPlayerGetPreferredOutputMode()` to improve handling of
   decode-to-texture mode.  For example, it now switch the output mode
   to punch-out for HDR videos, even if the requested mode is
   decode-to-texture.
5. Added `player_get_preferred_output_mode_test.cc` specific to Android
   TV, to verify the special handling on Android TV.

b/261922568

(cherry picked from commit 9f57b0b032dac27480433ab7f147e442fc6ec8c9)