[media] Workaround to support video poster
This CL allows the web app to set a poster image for video element by:
1. Playing the video in decode-to-texture mode.
2. Setting the poster image as the background image of the video element.

Previously the video background is painted to black when there are no
frames to display.  This isn't the expected behavior as it overwrites
any background image or color settings on the video element.

Now by default Cobalt won't paint the video background to black.

H5vcc setting `MediaElement.PaintingVideoBackgroundToBlack` is also
introduced so the web app can opt for the previous behavior, i.e. set
the background to black.

b/261922568