Cherry pick PR #807: [android] Reset frame rate estimator when the stream fps changed (#807) (#1062)
Cherry pick PR #807: [android] Reset frame rate estimator when the
stream fps changed (#807)

Refer to the original PR: https://github.com/youtube/cobalt/pull/807

When playing a stream, the key operating rate is set to MediaCodec for
resource planning. It's calculated by the playback rate and the
estimated fps. While the latter is calculated by the number of rendered
frames in duration. When the stream fps is changed, the estimated fps
will be affected by the frames already rendered with the old fps and
slowly adjusted to the new fps. This may lead to insufficient resources
prepared for the stream and cause some performance issues.

b/290409989

Change-Id: I5126794d0ac98d29fc8405359bcb4f62dc444308