Introduce sb_is_modular build flag (#1111)
b/246854012

sb_is_modular is set for modular builds, which have:
1) Application binary built as a shared library.
2) Either of a) or b)
a. Starboard built at a shared library and a separate loader_app
executable.
   b. A loader_app executable with Starboard built in ( Evergreen ).

sb_is_modular shares a large subset with evergreen but is different
because of certain elements which are evergreen specific. An example of
evergreen specific code not part of modular builds is
a)Cobalt app updater(cobalt/updater/updater.cc)

This change replaces certain common instances of sb_is_evergreen with
sb_is_modular
and SB_IS_EVERGREEN with SB_IS_MODULAR