[android] Replace SB_CHECK() with SB_LOG(WARNING) (#1152)
We refined the calculation of the dimensions of the decode target, and
left the SB_CHECK() to ensure that the transform matrix provided by the
platform produces the same result as the crop rectangles.

The SB_CHECK() was triggered on some devices when playing 360 videos.
Now it's replaced by SB_LOG(WARNING) to avoid the crash. The current
implementation falls back to the old logic automatically in this case so
360 videos continue to play.

b/294455449

(cherry picked from commit 158fdaee706a416a77333602e73b0d478b261311)