Make musl_unittests a static library, add a compiler assert checking sizeof(wchar_t) (#1263)
b/246854012

Add a compiler assert checking the size of of wchar_t == 2 bytes when
SB_IS_WCHAR_T_UTF16 is defined

musl_unittests is a compile time test, having it as a static library is
sufficient.
Making it a a shared library on certain windows-based platforms forces
it do define SbEventHandle - it's not defined currently and doesn't need
to be as these are just compile time checks which can be done as a
static library.
This is similar to the cpp17 audit test which has a bunch of similar
compile asserts - it's a static library

See b/246854012#comment55 for details.

(cherry picked from commit 0ad73d67bd9f99b54d06c808b5e92a6c403d1d3b)