Add RenderProcessHostObserver::InProcessRendererExiting
This new notification is useful for worker code that needs to observe
the render process exiting, which doesn't work in --single-process
mode.

Used in this CL to fix an issue where 2 RenderProcessHostObservers
were watching for the RenderProcessHostDestroyed() notification to
clean up nodes on the performance manager graph, but one of the
observer needed to be notified first, which depended on the somewhat
random order of the observers in the observer list.

This change is backported from m118+.

(cherry picked from commit 3426bf6e8f2eb618d09608b4379ee73f2c4c31a9)

Bug: 391938746
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4780551