[android] Refine audio output device querying (#1399)
`SbMediaGetAudioOutputCount()` was hard coded to return 1. This worked
when `SbMediaGetAudioConfiguration()` was hard coded to return HDMI, and
breaks 5.1 playbacks on some devices after
`SbMediaGetAudioConfiguration()` is refined to return configurations for
all audio outputs (including non-HDMI devices) in the system.

For example, when the first output is non-HDMI stereo, and the second
output is HDMI with 5.1 support, the second output won't be checked in
`IsAudioOutputSupported()` because `SbMediaGetAudioOutputCount()`
returns 1, and 5.1 streams will be incorrectly rejected.

Now `SbMediaGetAudioOutputCount()` tentatively returns 16 on Android TV,
and `IsAudioOutputSupported()` will break the loop when
`SbMediaGetAudioConfiguration()` returns false. This minimizes the
changes required while still allows 5.1 playbacks.

Follow up changes should be made to remove
`SbMediaGetAudioOutputCount()` completely once this is verified on all
platforms.

b/284140486
b/297426689