Fix OOB write in BuildHuffmanTable.
First, BuildHuffmanTable is called to check if the data is valid.
If it is and the table is not big enough, more memory is allocated.

This will make sure that valid (but unoptimized because of unbalanced
codes) streams are still decodable.

(cherry picked from commit 902bc9190331343b2017211debcec8d2ab87e17a)

Bug: chromium:1479274