Remove Service worker persistent setting origin url check
When register a service worker, we never check anything about the initial url. However,
in getRegistration(), it queries service worker persistent settings, which compares the
origin with the initial url. As a result, we can register a service worker, but never
get them from persistent setting file. And those service workers will not be unregistered
as well.

In addition to the fix here, we probably needs to have a better way to prevent the
Service worker persistent setting file to grow infinitely.

b/300293281