Add SetTimeZone API as a starboard extension
In order to test SbTimeZoneGetCurrent(), we need to set a time zone
first. Here we add a new time zone API as a starboard extension to
set time zone. We start with Linux implementation. Will have follow
up PR to add implementations for all first party platforms. This will
help catch the issue if a platform forget to implement
SbTimeZoneGetCurrent() which always return 0.

b/300108997