Enable platform-specific sources to compile loader
The loader app (for modular toolchain builds) currently only relies on
starboard_loader.cc as the source. For NX, we require additional
sources, and those can be provided by setting the value of the newly
added variable extra_platform_loader_sources in the configuration.gni
file for that platform.

b/246855300