Add 'memoryLimit' attribute to source_buffer.
This allows a web app to override the default memory_limit on a
SourceBufferStream which backs an instance of source_buffer.

A SourceBufferStream is not created until after the first Init Segment
is appended, so if you try to read the memory_limit before then, the web
app will get an exception. However you can still pre-emptively set the
memory_limit which will be applied to SourceBufferStream upon creation.

b/291291120

Change-Id: Id577df24fadad4bb6405a1ecf1439942cd988837