Add link seam example w/ Starboard proxy
This is an implementation of an idea raised by jfoks@ in a comment in
go/starboard-test-doubles. The Starboard proxy can be used to
register/unregister a test double to be used in place of a real
Starboard API; the test double can be used by the code under test, and
the platform's real implementation can be used by surrounding code in
the loader, googletest, etc.

This PR is just uploaded to gather feedback and won't be merged.

b/36360851

Change-Id: I786759517d954aae34707830be8efc962c956b2b