[android] Signal application readyness precisely. (#1711)
This signals precisely to the StarboardBridge when the
ApplicationAndroid is ready, by calling from the constructor and
destructor, and setting or clearing a new starboardApplicationReady
flag, used to gate all calls to native methods that can use
ApplicationAndroid, to ensure such calls are not made before the
constructor finishes or after the destructor starts.

The 'applicationStopped' flag is now only used to exit in
onActivityDestroy.

b/301158281

(cherry picked from commit c28f1850df609328c1b123635ba72d96a4a90418)