Add build action of RDK platform to Evergreen
RDK shares the evergreen-arm-hardfp with raspi-2 but since it has
its own pre-installed loader, it doesn't require a loader to be
built. Also it supports SB14 only.
For now it shares the build with raspi-2. Eventually it will
have its own build entry, and this task is logged as b/307558214.

b/293172465

Change-Id: Ia241ada9903d3b2ba427ad426ba02a89e7ccc94a