Fix telemetry native crash
Instead of storing a reference to a JS callback in the
metrics_log_uploader, we instead publish an event to EventDispatcher
when metrics are ready to be uploaded. H5vccMetrics listens for these
events and returns the result to the JS callback.

This eliminates race conditions where the JS environment (and therefore
H5vcc) was being destroyed and metrics were queued for upload,
referencing freed pointers.

b/298050585

Change-Id: I261436a00099d7640791bc3a9ec93aab600ed3db