Fix telemetry native crash (#1783)
Instead of storing a reference to a JS callback in the
metrics_log_uploader, we instead publish an event to EventDispatcher
when metrics are ready to be uploaded. H5vccMetrics listens for these
events and returns the result to the JS callback.

This eliminates race conditions where the JS environment (and therefore
H5vcc) was being destroyed and metrics were queued for upload,
referencing freed pointers.

b/298050585

(cherry picked from commit d9fdfb83e0472aeaadb5d6ba238798021f9a2f67)