[GN] Pass shared library name to exe install steps
When running the install steps for executables that are built as a
result of the shared library (i.e. loaders), sometimes the name of the
original shared library that invoked it is needed in the install_target
context in order to create necessary dependencies on the shared_library
install_target counterparts.

b/246855300