[NX] Fix bundle content steps for modular builds
Bundle content steps need to copy files from both cobalt and starboard
toolchain versions of the `content/data` directory. However to preserve
the folder structure so that we get the correct placement of
content/data files in the final packaged NX application, we need to
ensure the copy-script uses the correct toolchain-specific base
directory.

b/246855300

Change-Id: Iaffced7832bfc157b7c3c0a6d600b58768977ca0