Fix data_deps in modular loader executable install (#1838)
Since each test target will create a corresponding install folder with
its data files installed to it, it is necessary that this is replicated
for modular-toolchain where the test is both the shared library and
loader executable. When deploying, we resort to the install folder for
the loader executable which is why we forward the data_deps.

b/246855300

Change-Id: I159de3558009fddf23f4e2471bf74b4ac5563c7d
(cherry picked from commit 2254f6ff192716c1c81e48eb2480ef99ad9f97a2)