Create MicCreationSuccessful UMA histogram
Adds a boolean enum histogram to track the frequency at which mic
usage is successful. When mic input is requested in
mediaDevices.getUserMedia(), either a success or failure callback
is run depending on whether a mic was able to be opened. This
histogram logs a success or failure item in each callback.

See http://go/cobalt-mic-uma-histogram for additional details.

b/306441161