Add the RTLD_DEEPBIND flag when dlopen-ing FFmpeg libraries.
Without this, the cast runtime when built out of chromium will segfault
in starboard's FFmpeg decoder. The issue is that chromium builds a newer
version of FFmpeg, and (without RTLD_DEEPBIND) some of the symbols in
FFmpeg get resolved as the chromium versions. This likely causes a
segfault due to struct size changes (in my case, the segfault was in
av_buffer_unref). See the bug for more details, including a full stack
trace.

Bug: b/313524200