[M114-LTS] Vulkan: Don't crash when glCopyTexImage2D redefines itself
M114 merge issues:
  renderer/vulkan/TextureVk.cpp:
    Confliciting comments

The Vulkan backend marks a level being redefined as such before doing
the copy.  If a single-level texture was being redefined, it releases it
so it can be immediately reallocated.  If the source of the copy is the
same texture, this causes a crash.

This can be properly supported by using a temp image to do the copy, but
that is not implemented in this change.

Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5200755
git-subtree-dir: third_party/angle
git-subtree-split: 3f6ffdb286bdfcb8be570714201dca6e1c9f2027