Refine RDK test_runner script
Copy test data to a dedicated folder on RDK, so that it could be
easily removed after the test is done or before the next test to
save disk space. It would be cumbersome if all the tests are
stored at individual folders at the root directory.

Also add a new tag to scan keywords list.

b/307631462

Change-Id: I643d703d4be25eb1236c25a4db00060a26f493c6