[nplb] Increase PartialAudio test segment size (#2093)
We may see performance issue if segment size is too small. A segment
size of 0.1 would ask the platform to decode 10x audio units. A segment
size of 0.3 makes more sense as it ask the platform to decode 4x audio
units, which could happen when we have partial audio units with playback
speed of 2x.

b/313782907