Disable MessagePumpTest
MessagePumpUIStarboard::Run() is never called because we are not
like a normal message pump where we loop until told to quit. We
are providing a MessagePump interface on top of an externally-owned
message pump. We want to exist and be able to schedule work, but
actually for (::) is owned by Starboard.

The test failed as intended as some NOTREACHED() being called.

b/298237650
b/316012657