[android] Handle mediacodec callback with a handler thread (#2044)
Now the MediaCodec callbacks and the activity lifecycle callbacks are
called on the main thread. When pausing YouTube, the MediaCodec
callbacks may be blocked until the lifecycle callbacks are done. This
may cause frame drops on a resource-limited device when the MediaCodec
callback to notify a newly decoded frame is blocked and we run out of
decoded frames. Running the MediaCodec callbacks on a handler thread
could avoid this problem.

Change-Id: I5ffdb1f5a582c3d01964b3f98c99d7aed211674a

b/316008643