Add prototype for differential Evergreen updates
This prototype enables the loader app to apply a bsdiff patch, found in
an installation slot, to the system image binary to generate a new
Cobalt binary at load time.

It also enables a command-line tool to be built to generate bsdiff
patches that are consistent with what the loader app expects. If we ever
actually implemented differential updates, we'd probably generate the
patches server-side and provide them to clients during updates.

Finally, it uses Chromium's implementation of bsdiff from courgette/.
It's modified a bit to reduce the number of depdendencies.

b/174699165