Improve Nested Flexboxes.
This fixes calculation for the main direction length and position of flex
items in horizontal flex containers that are themselves also flex containers
in vertical direction.

This also improves the determination of whether a cross size is definite,
taking force_fit_to_shrink_width into account, as well as storing the
fit_to_shrink_width for both horizontal and vertical flex containers.

Tests will be added in a separate commit.

b/197582262