Remove the option to start crash handler at launch
This is a tech debt cleanup. Starting the crash handler when a crash
happens instead of when the app starts has been the default behavior.
This behavior saves 2-3MB memory that the app consumes. This behavior
has been proved reliable. The code to allow the other behavior can be
removed.

Change-Id: Icc725cb801a72b4635d6104abc8226a6f6323f46