Ensure posix_emu includes come before system includes (#2443)
We'll need to define emulated versions of header files which already
exist in the system, so we need our version of the file to be found
first. This also removes some no-op include_dirs sprinkled in a separate
config to help simplify our includes.

The initial use case is to have the Windows `open` API be available when
common POSIX code does `#include <fcntl.h>`. So, we'll make our own
fcntl.h that includes the system-level one, and also includes the
Windows-specific io.h that defines the `open` API.

Also cleans up some of the other posix_emu code to be consistent.

b/302715109

Test-On-Device: true