Make Visual Studio 2022 install Docker-Friendly (#2500)
This script introduces extra logic to check that the file exists when
deleting the cached and temporary files during installation. This would
break otherwise if the folders did not match exactly and that causes the
current shell environment to exit early.

Current the docker version of the visual studio 2022 installation
involves running an unwieldy and hard to read command that combines both
downloading the file, installing it, and deleting unnecessary folders
and caches.

Moving it into a single script invocation helps with constraining image
sizes. All steps need to happen sequentially in the same RUN directive
to avoid the final image size becoming the maximum of all intermediate
images from the RUN directives.

That is to say, 3 RUN directives (download, install, delete) will create
3 layers each with its own size, and the largest of the three affects
the final image size. Doing it atomically ensures only the final and
smallest sized layer is recorded and used.

b/326639379

Change-Id: I661fa2ce7b3e61c0c544ef6bfac2de6c54010fd2