Stabilize JavaScript Profiler in Cobalt (#2454)
b/323983545

A major refactor was needed in order to prevent the JavaScript Profiler
from SEGFAULTING due to a garbage-collected nullptr.

This refactor includes the use of a `ProfilerGroup` class. The life of a
cobalt::js_profiler::Profiler is extended by adding scoped_refptrs to
the ProfilerGroup; thus preventing GC. Profilers live as long as their
ProfilerGroup, which lives as long as the Isolate and Web Agent. One
`ProfilerGroup` is bound to one `web::Agent`, and contains one
`v8::Isolate`, one `v8::CpuProfiler`, and many
`cobalt::js_profiler::Profiler`s.

The major advantage of this refactor is that it does not crash, and is
more faithful to the [original Chromium
Implementation](https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/timing/profiler_group.cc;bpv=0;bpt=0).

Test-On-Device: true

---------

Co-authored-by: Ahmed Elzeiny <ahmedelzeiny@google.com>