Fix UploadFileElementReaderTest
In this PR: https://github.com/youtube/cobalt/pull/2252, we add
delete_on_close_ to fix file_unittest.cc. However, I don't think
that's the right approach and it breaks UploadFileElementReaderTest.
Currently, those FileTest is disabled in trunk. Revert that change
and leave it as it is for now.

b/327654843