Add telemetry for `SourceBuffer.appendBuffer`
Modifies the SourceBufferMetrics class to not be completely stubbed out
in gold builds, and use it to send UMA histograms for the time spent in
`SourceBuffer::PrepareAppend` and
`SourceBufferAppendAlgorithm::Process`.

Added basic unit tests for the `SourceBufferMetrics` class, to assert
that the new code reports telemetry against the correct histograms.

b/329439818

Change-Id: I7921434b02c689b645ab2a60464eee166d829135