Make loader app withstand failed Crashpad db init (#2598)
Because of a missed nullptr check, the loader app currently crashes if
initialization of the Crashpad database fails during installation of the
Crashpad handler. There will be further investigation and improvements
here but for now, the loader app will continue to load Cobalt - just
without crash reporting for the current session.

We shoud add unit tests for this module soon. It would probably be best
to stub out the Starboard calls, which will require a little
refactoring.

b/329458881
b/326436503

Change-Id: Ibb9e9abe5e0295ddbc6d139f11d4ab6b3a1fbb4c