Add a RunLoop to wrap_main_starboard
We used to create a message loop in wrap_main_starboard. Now that
message loop is gone and we create only a MessagePumpUIStarboard in
wrap_main_starboard, we need an additional RunLoop instance to execute
next tasks. I'm putting it in wrap_main_starboard instead of the
MessagePump because test::TaskEnvironment has an assumption that
MessagePump can be created before TaskRunner on each thread which means
MessagePump can not own a run_loop class since run_loop needs to know
which TaskRunner instance to ask for work at init.

Change-Id: I763bf89e38d18f01eb8b047e05aa6ca8249c3220