JavaScript Profiler Black box tests (#2569)
b/326457979

This set of BlackBox tests causes Cobalt to SEGFAULT by forcing Garbage
Collection to execute while the JS Profiler is running. It adds a new
API to CrashLog called h5vcc.crashLog.forceGarbageCollection() for the
purposes of testing. Reverts [an existing revert in PR
2567](https://github.com/youtube/cobalt/pull/2567). Tests are rewritten
to be less CPU-intensive; expecting fewer stack frames over a longer
period of time.

Test-On-Device: true

---------

Co-authored-by: Ahmed Elzeiny <ahmedelzeiny@google.com>