Properly cleanup HardwareBackendImage
In previous Cobalt releases, we depended on the rasterizer thread to
execute all pending tasks before exiting to delete a raw pointer of
HardwareBackendImage. This was dangerous and caused memory leak when
using the new base. Here we bind a smart pointer to the task so even
if the task is not executed the memory will be released.

b/333556585

Change-Id: I6c3056809ab3f5af099676b01cd440a38e298b8d