Apply Cobalt customizations to updated Chromium media (#2854)
This change modifies the `//media` directory to allow it to build
with Cobalt. These modifications are guarded with `#if
defined(STARBOARD)`.

This change also includes the following changes:

1. Applies upstream Chromium changes to support IAMF audio. These
changes are not guarded.
2. Creates new build targets for `//ui/gfx` and includes them in the
build.

Remaining gaps:
- MediaLog is stubbed out. We should remove the customization and enable
it. Tracked in b/230916218
- //skia and //ui/gfx have minor customizations that allow them to build
with Cobalt. Most or all of these should be removed once
//third_party/skia is updated to match Chromium.
- `PipelineStatusCB` is planned to be replaced with
`PipelineStatusCallback`. We should update Cobalt media to use only
`PipelineStatusCallback`. Tracked in b/335673619

b/319697087